Fix binding privileged ports as consul user
Listening on TCP/UDP ports below 1024 require to be the root user or to have the CAP_NET_BIND_SERVICE Linux capability.
Currently Consul is run as the consul user. As an example when a privileged port is configured for DNS, the following
error is reported upon container start:
==> Starting Consul agent...
==> Starting Consul agent RPC...
==> Error starting dns server: dns udp setup failed: listen udp 0.0.0.0:53: bind: permission denied

This commit grants the CAP_NET_BIND_SERVICE capability to the /bin/consul binary.