Add CONSUL_ADVERTISE_INTERFACE to supply advertise interface
Reference: hashicorp/docker-consul#59

This adds CONSUL_ADVERTISE_INTERFACE and related variables, which can be
used to pass in an advertise interface, versus a bind interface. This
allows you to listen on all interfaces in a --net=host situation (so
Consul listens on localhost as well), while ensuring that you can elect
to advertise on a specific interface to ensure that startup still
succeeds.